+++ R/facet_plot.R
@@ -3,8 +3,9 @@
 ##'
 ##' 'facet_plot()' automatically re-arranges the input 'data' according to the tree structure,
 ##' visualizes the 'data' on specific 'panel' using the 'geom' function with aesthetic 'mapping' and other parameters,
-##' and align the graph with the tree 'p' side by side.
+##' and align the graph with the tree 'p' side by side. 'geom_facet' is a 'ggplot2' layer version of 'facet_plot'
 ##' @title facet_plot
+##' @rdname facet-plot
 ##' @param p tree view
 ##' @param panel panel name for plot of input data
 ##' @param data data to plot by 'geom', first column should be matched with tip label of tree
@@ -20,85 +21,18 @@
 ##' @export
 ##' @author Guangchuang Yu
 facet_plot <- function(p, panel, data, geom, mapping=NULL, ...) {
-    p <- add_panel(p, panel)
-    df <- p %+>% data
-    p + geom(data=df, mapping=mapping, ...)
+    p + geom_facet(panel = panel, data = data,
+                   geom = geom, mapping = mapping, ...)
 }
 
-
-##' @importFrom ggplot2 facet_grid
-add_panel <- function(p, panel) {
-    df <- p$data
-    if (is.null(df[[".panel"]])) {
-        df[[".panel"]] <- factor("Tree")
-    }
-    levels(df$.panel) %<>% c(., panel)
-    p$data <- df
-    p + facet_grid(.~.panel, scales="free_x")
-}
-
-
-##' label facet_plot output
-##'
-##' 
-##' @title facet_labeller
-##' @param p facet_plot output
-##' @param label labels of facet panels
-##' @return ggplot object
-##' @importFrom ggplot2 labeller
+##' @rdname facet-plot
 ##' @export
-##' @author Guangchuang Yu
-facet_labeller <- function(p, label) {
-    ## .panel <- panel_col_var(p)
-    ## lbs <- panel_col_levels(p)
-    lbs  <-  levels(p$data$.panel)
-    names(lbs)  <-  lbs
-    label <- label[names(label) %in% lbs]
-    lbs[names(label)]  <-  label
-
-    ## ff <- as.formula(paste(" . ~ ", .panel))
-    p + facet_grid(. ~ .panel, scales="free_x",
-                   labeller = labeller(.panel = lbs))
+geom_facet <- function(panel, data, geom, mapping=NULL, ...) {
+    params <- list(...)
+    structure(list(panel = panel, data = data,
+                   geom = geom, mapping = mapping,
+                   params = params), class = 'facet_plot')
 }
 
 
-##' set relative widths (for column only) of facet plots
-##'
-##' 
-##' @title facet_widths
-##' @param p ggplot or ggtree object
-##' @param widths relative widths of facet panels
-##' @return ggplot object by redrawing the figure (not a modified version of input object)
-##' @author Guangchuang Yu
-##' @export
-##' @importFrom ggplot2 ggplot_gtable
-facet_widths <- function(p, widths) {
-    if (!is.null(names(widths))) {
-        ## if (is.ggtree(p) && !is.null(names(widths))) {
-        ## .panel <- levels(p$data$.panel)
-        .panel <- panel_col_levels(p)
-        w <- rep(1, length=length(.panel))
-        names(w) <- .panel
-        w[names(widths)] <- widths
-        widths <- w
-    }
-    gt  <- ggplot_gtable(ggplot_build(p))
-    for(i in seq_along(widths)) {
-        j <- gt$layout$l[grep(paste0('panel-', i), gt$layout$name)]
-        gt$widths[j] = widths[i] * gt$widths[j]
-    }
-    return(ggplotify::as.ggplot(gt))
-}
 
-panel_col_var <- function(p) {
-    m <- p$facet$params$cols[[1]]
-    if (is.null(m))
-        return(m)
-
-    ## rlang::quo_name(m)
-    rlang::quo_text(m)
-}
-
-panel_col_levels <- function(p) {
-    levels(p$data[[panel_col_var(p)]])
-}